add bannerWrapper component in Drive explorer instead of Banner component
